requirements.txt
boto3
openai
langchain
langgraph
llama-index
llama-index-vector-stores-chroma
llama-index-llms-openai
embedchain
chromadb
simplejson
streamlit
python-dotenv
watchdog
neo4j
transformers
py2neo
ollama
aura-cli
langchain-experimental
duckduckgo-search